+++ src/components/@common/Header/index.tsx
@@ -1,7 +1,7 @@
 import Svg from 'components/@common/Svg';
 import { useNavigate } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
-import { countCartListSelector } from 'recoil/cartList';
+import { countCartListSelector } from 'recoil/carts';
 import { ROUTES } from 'utils/constants';
 import SelectServer from '../SelectServer';
 import * as S from './Header.styles';
